fix checkbox
