--- frontend/src/validation/registerUserSchema.js
@@ -1,18 +1,0 @@
-import * as yup from 'yup'
-
-const registerSchema = yup.object().shape({
-  name: yup.string().required('this field is required'),
-  lastname: yup.string().required('this field is required'),
-  email: yup.string().email('must be a valid email').required('email is required'),
-  password: yup
-    .string()
-    .required('No password provided.')
-    .min(6, 'Password is too short - should be 6 chars minimum.')
-    .matches(
-      /^(?=.*[A-Za-z])(?=.*\d)(?=.*[@$!%*#?&])[A-Za-z\d@$!%*#?&]{6,}$/,
-      'Debe contener un carácter especial (@ $ ! % * # ? &) y al menos un número'
-    ),
-  privacy: yup.boolean().oneOf([true], 'You must accept the terms and conditions'),
-})
-
-export default registerSchema
